Create Summarize-SIS-Imports-Log
Python program to summarize the contents of the SIS Import Log file on the Canvas LMS. Reads a local parameters file to obtain the URL of the log file and access token for HTTP GET operations. Gets the file, walks through the contents, and displays only selected detaials. If the log file referenes one or more error files, each error file is also retrieved and the number of entries for each file type is summarized.