Create Summarize Canvas SIS Import Log
Summarize an SIS Imports Log File from the Canvas Learning Management System.

The program retrieves a log file from the web and extracts selected details from it. If an error or warnings file is referenced, that file is also retrieved and a count made of the number of lines referencing each file type. 

The program requires a customer-specific URL, and an access token that was generated within Canvas. These are read from a parameters file. The file is coded just so these keys are not listed in plain text. Since the decoding scheme is explicitly stated in the program code, the paramters file cannot be fully secured.