Create Run-and-Retrieve-Canvas-Report
This Python 3.8 program uses HTTP POST and GET to request the execution of a pre-defined report in the Canvas Learning Management System, to wait for the report to finish, then to retrieve the file containing the report output. The HTTP POST tells Canvas to start a report by a report name defined in Canvas, and passes report parameters such as term, start date, and course enrollment types. Program written by Jim Dugan on Windows 10, September 2020.