Create List-Canvas-Reports
Program lists the standard Canvas reports that are available through the User Interface at Admin > Settings > Reports. Report names are listed in the form used by Canvas APIs to  list specific instances of a report or delete a specific instance of a report.  Python requests.get() executes the HTTP call to retrieve the list. The program requires a URL that refers to a specific instance of Canvas, and an access token created within that instance of Canvas.