Create List-Report-Instances
Python program for Canvas API Call to list instances of a specific type of standard report. The standard reports are available through the User Interface at Admin > Settings > Reports. The API call implemented here is called "Index of Reports" in Canvas API documentation, and described as GET /api/v1/accounts/:account_id/reports/:report, with curl code equivalent:
     curl -H 'Authorization: Bearer <token>' \
Python requests.get() issues the HTTP GET.
     https://<canvas>/api/v1/accounts/<account_id>/reports/<report_type>
The reportType variable in the python code must have the name of a particular type of standard report as known to Canvas report APIs. A list of the names can be obtained for the List Available Reports API, which is implement in another module named List Canvas Reports, here in this GitHub repository. 

The program requires the URL of a specific instance of Canvas, and an access token created within that instance of Canvas. The program retrieves these from an encoded file.