Create Delete-a-Canvas-Report
Implements a Canvas API call to delete an instance of a standard Canvas report. Standard Canvas reports are available through the User Interface at Admin > Settings > Reports. The API implemented here is called "Delete a Report" in Canvas API documentation, define as AccountReportsController#destroy
DELETE /api/v1/accounts/:account_id/reports/:report/:id, with equivalent curl code:

curl -H 'Authorization: Bearer <token>' \
     -X DELETE \
     https://<canvas>/api/v1/accounts/<account_id>/reports/<report_type>/<id>

Python requests.delete() implements the HTTP DELETE.  The program needs a report type, which is the name of one of the types of standard reports, as known to the Canvas APIs. The program also requries the ID of a particular instance of that report type. API-compatible names of report types can be obtained through the "List Available Reports" API, and a list of specific instances of a report type can be obtained from the "Index of Reports" API.